+++ src/app/fork-me/fork-me.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-fork-me',
+  templateUrl: './fork-me.component.html',
+  styleUrls: ['./fork-me.component.scss']
+})
+export class ForkMeComponent implements OnInit {
+
+  constructor() {
+  }
+
+  ngOnInit() {
+  }
+
+}