+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+<a name="5.0.0"></a>
+# 5.0.0 (Feb 28, 2018)
+
+### Bug Fixes
+* Fix a non-issued error on modal `escapable` (previously `escapeAble`, c.f. Features below) option that wasn't closing the modal on escape event
+* Fix a non-issued error on modal `escapable` event that was called in each modal even if not visible. Now it only calls in the visible modals
+
+### Features
+* **BREAKING CHANGES**: rename `escapeAble` option to `escapable`
+* Create default z-index and check all modals (pull #79)
+* Add issue template (pull #80)
+* Add allowOutsideClick option (pull #84)
+* Rename `allowOutsideClick` option from PR #84 to `dismissable` for more consistency
+* Update unit tests
+* Update demo dependencies
+* Update the demo with `escapable` and `dismissable` options examples
+
+
 <a name="4.1.0"></a>
 # 4.1.0 (Dec 26, 2017)